# sample-service/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  description: "Manage static/mostly static ndb data"
  title: Static Data
# basePath prefixes all resource paths 
basePath: /api/v1
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /neurons/{neuronId}:
    x-swagger-router-controller: neuron
    get:
      summary: Get neuron definition
      tags:
        - Neurons
      operationId: getNeuronById
      parameters:
        - name: neuronId
          in: path
          description: Neuron object to find
          required: true
          type: string
      responses:
        200:
          description: The Neuron
          schema:
            $ref: "#/definitions/Neuron"
        default:
          description: Error
          schema:
            $ref: "#/definitions/Error"
  /samples/{sampleId}/neurons:
    x-swagger-router-controller: neuron
    get:
      summary: Neurons for a sample
      operationId: getNeuronsForSample
      tags:
        - Samples
        - Neurons
      parameters:
        - name: sampleId
          in: path
          description: Sample id for neuron lookup
          required: true
          type: string
      responses:
        200:
          description: An array of neurons for the given sample
          schema:
            type: array
            items:
              $ref: '#/definitions/Neuron'
        default:
          description: Error
          schema:
            $ref: "#/definitions/Error"
...
  /transforms/{transformId}:
    x-swagger-router-controller: registrations
    get:
      summary: Transform by id
      tags:
        - Registration Transforms
      operationId: getTransformById
      parameters:
        - name: transformId
          in: path
          description: Registration transform to find
          required: true
          type: string
      responses:
        200:
          description: The registration transform
          schema:
            $ref: "#/definitions/RegistrationTransform"
        default:
          description: Error
          schema:
            $ref: "#/definitions/Error"
...
definitions:
...
  Neuron:
    type: object
    properties:
      id:
        type: string
        description: Internal id
      sampleId:
        type: string
        description: Sample reference id
      createdAt:
        type: string
        format: date-time
        description: Creation time
      updatedAt:
        type: string
        format: date-time
        description: Update time
...
  RegistrationTransform:
    type: object
    properties:
      id:
        type: string
        description: Internal id
      name:
        type: string
        description: Name
      mutable:
        type: boolean
        description: Mutability
      createdAt:
        type: string
        format: date-time
        description: Creation time
      updatedAt:
        type: string
        format: date-time
        description: Update time
...
  Error: 
    type: object
    required:
      - message
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string